Fix regexp matching of image url in CSS image-set
The regular expression was too eager, so it was taking the entire image
set in its match rather than only a single one of the urls. As a result,
each time it ran rather than skipping the image, it would generate a new
even longer image-set, until it eventually hits the max string length
and breaks.

Fixes #3