Block loading of not already complete images
This explicit setting of src= to empty is needed to convince Chrome to
cancel loading of images. It prevents the problem where the square
thumbnail was initially showing up, then the fixed one appeared later. I
think it might actually speed up loading the fixed thumbnails, too.

Setting src= to empty is only done on images where !complete. If an
image was complete, then it's already shown - unsetting src would cause
a flash where the image disappears and reappears. This problem is particularly
noticable on pages with staticly loaded images, like the bookmark edit
page, but also might happen if the original image was in cache.